PUSH 08.01.2019 LOL SHOP - PAY STARTED

+++ Tankstelle/BusinessLogic/BusinessLogic/Product.cs
@@ -10,7 +10,7 @@ namespace BusinessLogic
     {
         public string Name { get; set; }
 
-        public double Price { get; set; }
+        public float Price { get; set; }
 
         public string PriceInString
         {
